src/main/java/com/active/jpa/service/OrderService.java
package com.active.jpa.service;

import com.active.jpa.domain.Delivery;
import com.active.jpa.domain.Member;
import com.active.jpa.domain.Order;
import com.active.jpa.domain.OrderItem;
import com.active.jpa.domain.item.Item;
import com.active.jpa.repository.ItemRepository;
import com.active.jpa.repository.MemberRepository;
import com.active.jpa.repository.OrderRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@Transactional(readOnly = true)
@RequiredArgsConstructor
public class OrderService {

    private final OrderRepository orderRepository;
    private final MemberRepository memberRepository;
    private final ItemRepository itemRepository;

    // 주문
    @Transactional
    public Long order(Long memberId, Long itemId, int count) {
        // 엔티티조회
        Member member = memberRepository.findOne(memberId);
        Item item = itemRepository.findOne(itemId);

        // 배송정보 생성
        Delivery delivery = new Delivery();
        delivery.setAddress(member.getAddress());

        // 주문상품 생성
        OrderItem orderItem = OrderItem.createOrderItem(item, item.getPrice(), count);

        // 주문 생성
        Order order = Order.createOrder(member, delivery, orderItem);

        // 주문 저장
        orderRepository.save(order);

        return order.getId();
    }

    @Transactional
    public void cancel(Long orderId) {
        Order order = orderRepository.findOne(orderId);
        order.cancel();
    }

//    @Transactional
//    public List<Order> findOrder(Ordersearch ordersearch) {
//
//    }
}
